requirements.txt
apache-airflow==2.9.0
apache-airflow-providers-snowflake==5.0.0
snowflake-connector-python==3.10.0
